CRUD Locação de Veiculos

+++ LocacaoVeiculo/src/main/java/com/G3/scm/controller/VeiculoController.java
@@ -1,19 +1,22 @@
 package com.G3.scm.controller;
+import javax.validation.ConstraintViolationException;
 import javax.validation.Valid;
+
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Controller;
 import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.client.HttpClientErrorException;
 import org.springframework.web.servlet.ModelAndView;
 
-
 import com.G3.scm.model.Veiculo;
-import com.G3.scm.servico.VeiculoServico;
+import com.G3.scm.servico.VeiculoServicoI;
 
 @Controller
 @RequestMapping(path = "/sig")
@@ -21,7 +24,7 @@ public class VeiculoController {
 	Logger logger = LogManager.getLogger(VeiculoController.class);
 	
 	@Autowired
-	VeiculoServico servico;
+	VeiculoServicoI servico;
 
 	@GetMapping("/menuVeiculo")
 	public ModelAndView menuVeiculo() {
@@ -51,11 +54,24 @@ public class VeiculoController {
 	
 	@GetMapping("/veiculo/{id}")
 	public ModelAndView excluirNoFormDeConsultaVeiculo(@PathVariable("id") Long id) {
-		servico.deleteById(id);
+		ModelAndView modelAndView = new ModelAndView("consultarVeiculo");
+		Veiculo veiculo = servico.findById(id);
+		if(veiculo.isLocado() == true) {
+			modelAndView.addObject("message", "Veiculo esta alocado nesse momento, impossivel deletar.");
+		}
+		else {
+			servico.deleteById(id);
+			logger.info(">>>>>> 1. servico de exclusao chamado para o id => " + id);
+		}
+		
+		
+		modelAndView.addObject("veiculos", servico.findAll());
+		return modelAndView;
+		/*servico.deleteById(id);
 		logger.info(">>>>>> 1. servico de exclusao chamado para o id => " + id);
 		ModelAndView modelAndView = new ModelAndView("consultarVeiculo");
 		modelAndView.addObject("veiculos", servico.findAll());
-		return modelAndView;
+		return modelAndView;*/
 	}
 	
 	@PostMapping("/veiculos")
@@ -64,9 +80,34 @@ public class VeiculoController {
 		if (result.hasErrors()) {
 			modelAndView.setViewName("cadastrarVeiculo");
 		} else {
-			modelAndView = servico.saveOrUpdate(veiculo);
+			try {
+				modelAndView = servico.save(veiculo);
+			}
+			catch (HttpClientErrorException e) {
+				modelAndView.addObject("message", "Dados invalidos - 400 Bad Request");
+				modelAndView.setViewName("cadastrarVeiculo");
+			} catch (ConstraintViolationException e) {
+				modelAndView.addObject("message", "Dados invalidos - constraint violation");
+				modelAndView.setViewName("cadastrarVeiculo");
+			} catch (DataIntegrityViolationException e) {
+				modelAndView.addObject("message", "Dados invalidos - veiculo já cadastrado");
+				modelAndView.setViewName("cadastrarVeiculo");
+			} catch (Exception e) {
+				modelAndView = new ModelAndView("cadastrarVeiculo");
+				modelAndView.addObject("message", "Erro não esperado - contate o administrador ==>" + e.getMessage());
+				logger.error(">>>>> 5. erro nao esperado ==> " + e.getMessage());
+				logger.error(">>>>> 5. erro nao esperado ==> " + e.toString());
+			}
+			
 		}
 		return modelAndView;
+		/*ModelAndView modelAndView = new ModelAndView("consultarVeiculo");
+		if (result.hasErrors()) {
+			modelAndView.setViewName("cadastrarVeiculo");
+		} else {
+			modelAndView = servico.saveOrUpdate(veiculo);
+		}
+		return modelAndView;*/
 	}
 	
 	@PostMapping("/veiculos/{id}")
@@ -79,11 +120,11 @@ public class VeiculoController {
 // programacao defensiva - deve-se verificar se o Cliente existe antes de atualizar
 		Veiculo umVeiculo = servico.findById(id);
 		umVeiculo.setPlaca(veiculo.getPlaca());
-		umVeiculo.setNome(veiculo.getNome());
+		//umVeiculo.setFabricante(veiculo.getFabricante());
 		umVeiculo.setModelo(veiculo.getModelo());
 		umVeiculo.setCor(veiculo.getCor());
 		umVeiculo.setValorDiaria(veiculo.getValorDiaria());
-		modelAndView = servico.saveOrUpdate(umVeiculo);
+		modelAndView = servico.save(umVeiculo);
 		return modelAndView;
 	}
 }
